update avatar card
