+++ package.json
@@ -47,7 +47,9 @@
     ]
   },
   "devDependencies": {
-    "@types/styled-components": "^5.1.25"
+    "@types/styled-components": "^5.1.25",
+    "jest-watch-typeahead": "^0.6.5",
+    "prettier": "2.8.8"
   },
   "packageManager": "yarn@3.5.0"
 }
