+++ src/components/AvatarCard/AvatarCard.styles.ts
@@ -1,0 +1,23 @@
+export const CardContainer ={
+    display: 'flex',
+    flexDirection: 'column',
+    alignItems: 'center',
+    maxWidth: 400,
+    borderRadius: 5,
+}
+
+export const Box = {
+    display: 'flex',
+    justifyContent: 'center',
+    width: '100%',
+    p: 2,
+}
+
+export const CardContent = {
+    p: 3, 
+    width: '100%'
+}
+
+export const Typography = { mb: 1, textAlign: 'center' }
+
+export const CardMedia = { maxWidth: 200 }
