+++ src/components/MarketplaceListing/MarketplaceListing.styles.ts
@@ -1,5 +1,3 @@
-import styled from 'styled-components'
-
 export const Container ={
     display: 'flex',
     marginTop: '100px',
